requirements.txt
einops==0.8.0
h5py==3.9.0
joblib==1.4.2
librosa==0.10.2.post1
lightning==2.3.3
lightning_utilities==0.11.6
matplotlib==3.9.1
numpy
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.1
optuna==3.6.1
pandas==2.2.2
rich==13.7.1
rootutils==1.0.7
scipy==1.14.0
setuptools==69.5.1
torch==2.3.0
torchaudio==2.3.0
torchmetrics==1.4.0.post0
tqdm==4.66.4
wandb==0.17.5
webdataset==0.2.86
hydra-core==1.3.2
hydra-colorlog==1.2.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
gdown
